CI: Fix Azure Pipelines macOS runs

+++ CI/install-dependencies-osx.sh
@@ -30,7 +30,9 @@ brew update
 #Base OBS Deps and ccache
 brew install jack speexdsp ccache mbedtls freetype fdk-aac
 brew install https://gist.githubusercontent.com/DDRBoxman/9c7a2b08933166f4b61ed9a44b242609/raw/ef4de6c587c6bd7f50210eccd5bd51ff08e6de13/qt.rb
-brew unlink swig
+if [ -d "$(brew --cellar)/swig" ]; then
+    brew unlink swig
+fi
 brew install https://gist.githubusercontent.com/DDRBoxman/4cada55c51803a2f963fa40ce55c9d3e/raw/572c67e908bfbc1bcb8c476ea77ea3935133f5b5/swig.rb
 
 pip install dmgbuild
